pyproject.toml
[tool.poetry]
name = "assignment"
version = "0.1.0"
description = ""
authors = ["Siddhesh Parave <siddeshparave16@gmail.com>"]
readme = "README.md"
packages = [
    { include = "api", from = "." }  # Critical change
]

[tool.poetry.dependencies]
python = ">=3.11"
django = "^5.2.3"
djangorestframework = "^3.16.0"
psycopg = "^3.2.9"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
django-phone-field = "^1.8.1"
djangorestframework-simplejwt = "^5.5.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"